refactor(panel-sync): правило даты переезжает в пакет синхронизации
Первый шаг консолидации: app/services/panel_expiry.py становится
app/services/panel_sync/expiry.py, у пакета появляется фасад. Поведение и
сигнатуры не менялись, переехал только адрес модуля.
